support for notification faces.extracted has been added
