+++ .gitignore
@@ -1,0 +1,30 @@
+*~
+*.user
+*.suo
+*.bak
+*.log
+*.sdf
+*.o
+*.ivec
+*.project
+*.ipch
+*.jfm
+*.dbmdl
+*.mp4
+build
+build_tmp
+[Dd]ebug
+[Rr]elease
+[Bb]in
+*.slite3
+*.lic
+*.pyc
+*_no_doc*
+build_logs
+changelog.html
+.idea/
+.vs/
+[Oo]bj
+_release
+packages
+Output/*
